requirements.txt
streamlit
Pillow
streamlit-drawable-canvas
opencv-python-headless